feat: 기존의 상품 목록 출력 코드를 각 모듈로 분리
- 현재는 목 데이터 사용 중으로 인터넷 연동 후 코드 변경
- app 모듈은 초기화만 담당하도록 변경 예정

- 컨벤션 모듈의 여러 플러그인에 누락된 임포트 문 추가

- complie -> compile 과 같은 오타 수정 포함
- jetbrations -> jetbrains
- ApplicationExtention -> ApplicationExtension

- configureKotlinAndroid에서 sdk 버전 번호 하드 코딩 제거

- compose 사용하는 모듈의 빌드 설정 파일에 컴파일러 플러그인을 추가함 (compose 관련 플러그인으로 분리 가능하면 리팩토링)

TODO:
- Room, Hilt 컨벤션 플러그인 추가
- feature 모듈 생성
