+++ app/build.gradle.kts
@@ -1,10 +1,10 @@
 plugins {
-    alias(libs.plugins.multi.module.android.application)
     alias(libs.plugins.multi.module.android.application.compose)
+    alias(libs.plugins.kotlin.compose)
 }
 
 android {
-    namespace = "com.multi.module.template"
+    namespace = "com.example.noti.template"
 
     defaultConfig {
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
@@ -12,12 +12,22 @@ android {
 }
 
 dependencies {
+    // hilt 도입 후 제거 예정
+    implementation(projects.home.presentation)
+    implementation(projects.core.presentation.designsystem)
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
+
+    implementation(libs.androidx.ui)
+    implementation(libs.androidx.ui.graphics)
+    implementation(libs.androidx.ui.tooling.preview)
+    implementation(libs.androidx.material3)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
     androidTestImplementation(libs.androidx.ui.test.junit4)
+    debugImplementation(libs.androidx.ui.tooling)
+    debugImplementation(libs.androidx.ui.test.manifest)
 }
