+++ app/src/main/java/com/example/noti/MainActivity.kt
@@ -4,33 +4,11 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Refresh
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.example.noti.ui.theme.NotiTheme
+import com.example.noti.core.presentation.designsystem.theme.NotiTheme
+import com.example.noti.home.presentation.HomeRootScreen
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -39,84 +17,9 @@ class MainActivity : ComponentActivity() {
         setContent {
             NotiTheme {
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(innerPadding)
-                    ) {
-                        SearchBar()
-                        CategorySection()
-                        ProductListScreen()
-                    }
+                    HomeRootScreen(paddingValues = innerPadding)
                 }
-
             }
         }
     }
 }
-
-@Preview(showSystemUi = true, name = "Full Screen with System UI")
-@Composable
-fun FullScreenWithSystemUI() {
-    NotiTheme {
-        Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(innerPadding)
-            ) {
-                SearchBar()
-                CategorySection()
-                ProductListScreen()
-            }
-        }
-    }
-}
-
-@Composable
-fun SearchBar() {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        horizontalArrangement = Arrangement.Start,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Icon(
-            imageVector = Icons.Default.Search,
-            contentDescription = "Search"
-        )
-        Text(text = "Search...", modifier = Modifier.padding(horizontal = 8.dp))
-        Icon(imageVector = Icons.Default.Refresh, contentDescription = "Refresh")
-    }
-}
-
-@Composable
-fun CategorySection() {
-    val categoryNames = arrayOf("Outer", "Top", "Pants", "Shoes", "Watch", "Bags")
-    Text(text = "Categoty")
-    LazyRow(
-        modifier = Modifier
-            .fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-        contentPadding = PaddingValues(8.dp)
-    ) {
-        items(categoryNames) { categoryName ->
-            CategoryButton(name = categoryName)
-        }
-    }
-}
-
-@Composable
-fun CategoryButton(name: String) {
-    Box(
-        modifier = Modifier
-            .size(60.dp)
-            .clip(CircleShape)
-            .background(Color.LightGray),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(text = name, color = Color.DarkGray)
-    }
-}
-
