--- app/src/main/java/com/example/noti/ProductData.kt
@@ -1,14 +1,0 @@
-package com.example.noti
-
-// 목 데이터
-val productList = listOf<Product>(
-    Product(1, R.drawable.shoes, "블랙 슈즈", 150000),
-    Product(2, R.drawable.cap, "블랙 캡", 55000),
-    Product(1, R.drawable.shoes, "블랙 슈즈", 150000),
-    Product(2, R.drawable.cap, "블랙 캡", 55000),
-    Product(5, R.drawable.top, "베이지 셔츠", 45000),
-    Product(1, R.drawable.shoes, "블랙 슈즈", 150000),
-    Product(5, R.drawable.top, "베이지 셔츠", 45000),
-    Product(2, R.drawable.cap, "블랙 캡", 55000),
-    Product(5, R.drawable.top, "베이지 셔츠", 45000),
-)
