+++ build-logic/convention/src/main/java/AndroidApplicationComposeConventionPlugin.kt
@@ -4,7 +4,7 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.getByType
 
-class AndroidApplicationComposeConventionPlugin: Plugin<Plugin> {
+class AndroidApplicationComposeConventionPlugin: Plugin<Project> {
     override fun apply(target: Project) {
         target.run {
             pluginManager.apply("multi.module.android.application")
