+++ build-logic/convention/src/main/java/AndroidLibraryConventionPlugin.kt
@@ -1,11 +1,14 @@
 /*  Compose가 없는 AndroidLibrary를 위한 컨벤션 플러그인 */
 
 import com.android.build.api.dsl.LibraryExtension
+import com.example.noti.convention.configureBuildTypes
 import com.example.noti.convention.configureKotlinAndroid
-import com.example.noti.convention.libs
+import com.example.noti.convention.ExtensionType
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
+import org.gradle.kotlin.dsl.dependencies
+import org.gradle.kotlin.dsl.kotlin
 
 class AndroidLibraryConventionPlugin: Plugin<Project> {
     override fun apply(target: Project) {
