+++ build-logic/convention/src/main/java/com/example/noti/convention/Extensions.kt
@@ -8,4 +8,4 @@ import org.gradle.kotlin.dsl.getByType
 /* Project.libs 라는 속성을 만들어, 어떤 Gradle Project 객체든 
 libs라는 이름으로 Version Catalog에 접근할 수 있도록 함. */
 val Project.libs: VersionCatalog
-    get() = extensions.getByType<VersionCatalogsExtension>.named("libs")
+    get() = extensions.getByType<VersionCatalogsExtension>().named("libs")
