+++ core/presentation/designsystem/build.gradle.kts
@@ -1,5 +1,7 @@
 plugins {
     alias(libs.plugins.multi.module.android.library.compose)
+//    alias(libs.plugins.multi.module.android.presentation.ui)
+    alias(libs.plugins.kotlin.compose)
 }
 
 android {
@@ -11,6 +13,10 @@ dependencies {
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
+    implementation(libs.androidx.ui)
+    implementation(libs.androidx.ui.graphics)
+    implementation(libs.androidx.ui.tooling.preview)
+    implementation(libs.androidx.material3)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
