+++ core/presentation/designsystem/src/main/java/com/example/noti/core/presentation/designsystem/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.example.noti.ui.theme
+package com.example.noti.core.presentation.designsystem.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
