--- home/domain/src/main/java/com/example/noti/home/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.example.noti.home.domain
-
-class MyClass {
-}
